backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
django-simple-history==3.10.1
djangorestframework==3.16.0
mysqlclient==2.2.7
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
